+++ src/helper/tweet.js
@@ -1,0 +1,16 @@
+const filterTweets = tweets =>
+	tweets.filter(_ => _.text.length > 5).filter(_ => !_.text.includes('t.co'));
+
+const filterTweetText = tweet =>
+	capitalizeFirstLetter(
+		tweet
+			.replace(/\n/g, ' ')
+			.replace(/\#/g, '')
+			.replace(/\@/g, '')
+			.trim()
+	);
+
+module.exports = {
+	filterTweetText,
+	filterTweets,
+};
